Move base SerDe to _singerlib
